requirements.txt
groq
pandas
numpy
scikit-learn
matplotlib
python-dotenv
